requirements.txt
pyserial
numpy
# picamera  # Uncomment if running on Pi with specific hardware
# opencv-python # Uncomment if using OpenCV strategy
